Update scala and sbt versions for scala-steward
This repository recently caused scala-steward to fail, because it
doesn’t currently meet the minimum required sbt and scala versions for
scala steward to work. (Those versions are outlined in [this
issue](https://github.com/guardian/maintaining-scala-projects/issues/7).

This commit updates the sbt version in use to the newest versions and
the two scala versions in use to the latest patch versions for those
minor versions.

This should be sufficient to let scala-steward run on this repository: I
have managed to successfully run `sbt +test` locally with these changes
using Java 21.